many compiler warnings fixes and some updates - merge Cruz Vincent's vidmode update

+++ display/X/box.c
@@ -1,4 +1,4 @@
-/* $Id: box.c,v 1.5 2003/06/12 12:09:39 cegger Exp $
+/* $Id: box.c,v 1.6 2003/07/06 10:25:21 cegger Exp $
 ******************************************************************************
 
    LibGGI - boxes for display-x
@@ -86,7 +86,7 @@ int GGI_X_drawbox_slave_draw(ggi_visual *vis, int x, int y, int w, int h)
 	y = GGI_X_WRITE_Y;
 	ggLock(priv->xliblock);
 	XFillRectangle(priv->disp, priv->drawable, priv->gc,
-			x, y, w, h);
+			x, y, (unsigned)w, (unsigned)h);
 	GGI_X_MAYBE_SYNC(vis);
 	ggUnlock(priv->xliblock);
 	return GGI_OK;
@@ -107,7 +107,7 @@ int GGI_X_copybox_slave_draw(ggi_visual *vis, int x, int y,
 	ny += LIBGGI_VIRTY(vis) * vis->w_frame_num;
 	ggLock(priv->xliblock);
 	XCopyArea(priv->disp,  priv->drawable, priv->drawable,
-		  priv->gc, x, y, w, h, nx, ny);
+		  priv->gc, x, y, (unsigned)w, (unsigned)h, nx, ny);
 	GGI_X_MAYBE_SYNC(vis);
 	ggUnlock(priv->xliblock);
 	return GGI_OK;
@@ -121,7 +121,7 @@ int GGI_X_drawbox_draw(ggi_visual *vis, int x, int y, int w, int h)
 	y = GGI_X_WRITE_Y;
 	ggLock(priv->xliblock);
 	XFillRectangle(priv->disp, priv->drawable, priv->gc,
-                       x, y, w, h);
+                       x, y, (unsigned)w, (unsigned)h);
 	GGI_X_MAYBE_SYNC(vis);
 	ggUnlock(priv->xliblock);
 	return GGI_OK;
@@ -136,8 +136,8 @@ int GGI_X_putbox_draw(ggi_visual *vis, int x, int y, int w, int h, void *data)
 	ggLock(priv->xliblock);
 #warning 1,2,4-bit support needed.
 	ximg = XCreateImage(priv->disp, priv->vilist[priv->viidx].vi->visual,
-			    LIBGGI_PIXFMT(vis)->depth, ZPixmap, 0,
-			    data, w, h, 8, 0);
+			    (unsigned)LIBGGI_PIXFMT(vis)->depth, ZPixmap, 0,
+			    data, (unsigned)w, (unsigned)h, 8, 0);
 	y = GGI_X_WRITE_Y;
 
 #ifdef GGI_LITTLE_ENDIAN
@@ -149,7 +149,7 @@ int GGI_X_putbox_draw(ggi_visual *vis, int x, int y, int w, int h, void *data)
 #endif
 
 	XPutImage(priv->disp, priv->drawable, priv->gc, ximg,
-		  0, 0, x, y, w, h);
+		  0, 0, x, y, (unsigned)w, (unsigned)h);
 	XFree(ximg); /* XDestroyImage would free(data) (bad).
 			Luckily, this doesn't leak (?) */
 	GGI_X_MAYBE_SYNC(vis);
@@ -167,7 +167,7 @@ int GGI_X_copybox_draw(ggi_visual *vis, int x, int y,
 	ny += LIBGGI_VIRTY(vis) * vis->w_frame_num;
 	ggLock(priv->xliblock);
 	XCopyArea(priv->disp,  priv->drawable, priv->drawable,
-		  priv->gc, x, y, w, h, nx, ny);
+		  priv->gc, x, y, (unsigned)w, (unsigned)h, nx, ny);
 	GGI_X_MAYBE_SYNC(vis);
 	ggUnlock(priv->xliblock);
 	return GGI_OK;
@@ -202,7 +202,7 @@ int GGI_X_getbox_draw(ggi_visual *vis, int x, int y, int w, int h, void *data)
 	/* This will cause a BadMatch error when the window is
 	   iconified or on another virtual screen... */
 	ximg = XGetImage(priv->disp, priv->drawable, x, GGI_X_READ_Y,
-			 w, h, AllPlanes, ZPixmap);
+			 (unsigned)w, (unsigned)h, AllPlanes, ZPixmap);
 	XSync(priv->disp, 0);
 	XSetErrorHandler(olderrorhandler);
 
@@ -258,7 +258,8 @@ int GGI_X_getbox_draw(ggi_visual *vis, int x, int y, int w, int h, void *data)
 		ximgptr = ximg->data + 
 			(ximg->xoffset * ximg->bits_per_pixel)/8;
 		while (h--) {
-			memcpy(data, ximgptr, (w * ximg->bits_per_pixel)/8);
+			memcpy(data, ximgptr,
+				(size_t)(w * ximg->bits_per_pixel)/8);
 			ximgptr += ximg->bytes_per_line;
 			((uint8 *)data) += 
 			  ximg->width * ximg->bits_per_pixel/8;
